backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
Pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
razorpay==2.0.0
redis==7.1.0
django-redis==5.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==14.0.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
whitenoise==6.11.0
sentry-sdk==2.19.2
django-storages==1.14.4
boto3==1.35.94
django-imagekit==5.0.0
django-axes==7.0.0
django-ratelimit==4.1.0
pyotp==2.9.0
qrcode==8.0
elasticsearch==9.2.0
django-elasticsearch-dsl==9.0
locust==2.33.0
